Add betting update functionality to BettingBoxController and related services

+++ KC.Backend.Logic/Interfaces/IPlayerLogic.cs
@@ -10,6 +10,6 @@ public interface IPlayerLogic
     void RemovePlayer(MacAddress playerId);
     Player Get(MacAddress playerId);
     void UpdateName(MacAddress playerId, string name);
-    void UpdateBalance(MacAddress playerId, int balance);
+    void UpdateBalance(MacAddress playerId, double balance);
     void UpdatePlayerConnectionId(MacAddress playerId, string connectionId);
 }
